Change input in Fibonacci function to short

+++ src/Fibonacci.cpp
@@ -1,6 +1,6 @@
 #include "Fibonacci.hpp"
 
-unsigned int Fibonacci(const unsigned int index)
+unsigned int Fibonacci(const unsigned short index)
 {
     return (index == 0 || index == 1)
            ? index
