+++ src/Fibonacci.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-unsigned int Fibonacci(unsigned int index);
+unsigned int Fibonacci(unsigned short index);
